Had many troubles with implementing csaps within this algo. Regardless of the validaty of the arrays passed to the csaps method, I would recieve an error that would have to do with the 'truth values of an array with more than one element...' which was an exception thrown up by a module within scipy that csaps must use.
The data that was being sent to csaps was an array of type float64 and
followed the only restriction within csaps (x1 < x2 < x3 < ... < xN).
Due to this roadblock and not much info online about possible fixes
an alternative spline approximation package was employed. Specifically
splrep is used from the scipy library.

The algo now passes tests.