diffraction calibration: reference-pixel selection
The first stage of the diffraction-calibration process involves the calculation and correction for the relative offsets of pixels in each detector group.  Ideally, each pixel should be measuring the same spectrum, although rescaled by a factor (DIFC) corresponding to the offset of that pixel.  Relative offsets  are calculated by taking the cross correlation of each pixel's spectrum with that of a reference pixel, and then calculating the shift of the cross-correlation peaks.  Since detector grouping assignment is arbitrary, it is important that the reference pixel be selected from within a detector group.  As it happens, this within-group selection is more important for the synthetic data used in the unit tests, than for data generated by the real instrument.  However, it should be noted that out-of-group selection might result in selection of a reference pixel currently assigned to _no_ group, in either case.

For synthetic test data used in the diffraction-calibration unit tests, the previous reference-pixel selection always resulted in out-of-group selection, as these groups included an _even_ number of pixels with non-consecutive workspace indices.  In addition, as each group only included detectors at exactly the _same_ physical offset, this caused significant changes in the behavior of the calibration process, in comparison to in-group reference pixel selection.  As noted previously, this should not be an issue on the real instrument.