diffraction calibration: xMin, xMax parameters
During the diffraction-calibration process, the ROI used by `GetDetectorOffsets` was previously hard coded to [-100, 100] bins.  This setting works in many cases.  However, depending on the d-space shift limit of the preceeding `CrossCorrelation` step, and the expected peak width, this setting doesn't _always_ work. The changes of this commit allow the `xMin`, and `xMax` parameters to scale with the cross-correlation d-space shift.

Note that the specific scale factor used was found empirically using the synthetic test data.  As such, it might require some adjustment.  In most cases, this scale factor should produce an ROI approximately the same as that used previously.