diffraction calibration: xMin, xMax parameters (#191)
* diffraction calibration: xMin, xMax parameters

During the diffraction-calibration process, the ROI used by `GetDetectorOffsets` was previously hard coded to [-100, 100] bins.  This setting works in many cases.  However, depending on the d-space shift limit of the preceeding `CrossCorrelation` step, and the expected peak width, this setting doesn't _always_ work. The changes of this commit allow the `xMin`, and `xMax` parameters to scale with the cross-correlation d-space shift.

Note that the specific scale factor used was found empirically using the synthetic test data.  As such, it might require some adjustment.  In most cases, this scale factor should produce an ROI approximately the same as that used previously.

* fixups to CIS-test script

* fix the issue with lite mode output (#196)

* fix the issue with lite mode output

* fix test and other algos

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* extra touchup to lite algo

---------

Co-authored-by: Boston S R <4rx@analysis-node18.sns.gov>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>

---------

Co-authored-by: Kort Travis <traviska@ornl.gov>
Co-authored-by: Reece Boston <52183986+rboston628@users.noreply.github.com>
Co-authored-by: Boston S R <4rx@analysis-node18.sns.gov>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>