Reduction output: effective instrument geometry.
At the last step of the reduction process, the real SNAP instrument is replaced with an effective instrument
consisting of one pixel per output spectrum.  The physical location of each of these effective pixels is
computed as the unmasked mean of the physical locations of original pixels in the group contributing
to the spectrum concerned.

Note that since SNAPRed's `PixelGroupingParameters` were already calculated as a mean of several
physical values over pixels contributing to each group, the required changes were implemented
by adding a few additional parameters to this structure.  Both masked and unmasked calculations are
available using the existing `GroceryService` and `SousChef` infrastructure.

This work consists of the following changes:

  * Add 'L2' and 'azimuth' to 'PixelGroupingParameters';

  * Clean up test code so that golden data can more easily be generated and compared.

  * Add an exception clause to deal with instabilities in `DetectorInfo.azimuthal()` which
  do not allow it to successfully complete the azimuth calculation for all pixels in the SNAP instrument.
  This will be spun off as a separate defect EWM#5073.