Ewm11197 calibration lookup by state (#595)
* This is intended to fix the way snapred reads calibration/normalization workspaces such that users do not require read perms to anything besides the reduction run and the calibration root.

Right now snapred requires read permissions to the original files used to generate the Calibration and Normalization files.
This is because it uses their logs to generate the state id we use for lookup, but this is redundant as we can generate that from the target sample run.
In addition, we already supply "version" as part of Calibration and Normalization lookup, which means at Service level we already have access to this info.

Thus, this pr refactors the calibration and normalization lookups to remove the optional `alternativeState`, and instead require `state`.

* fix broken unit tests

* delete comments