extensible state-ID schema
When the configuration of the SNAP instrument changes in a manner that affects its measurement results, this configuration defines a new calibration state.  Such a state can be identified using a subset of the PV-log values, which are usually rounded somewhat to remove artifacts such as encoder jitter.  This same set of PV-log values, though unrounded, is also used to calculate derived quantities which affect reduction results.

In SNAPRed, the calibration state is succinctly represented by a SHA-generated hash value.  This hash value is quite important, as it not only identifies a state, but it is also used to classify calibration and reduction data associated with that state.  Previously, the selection of PVs used to generate the calibration state was hard coded; the SHA-generated hash steps were hard coded; and any values derived from the state PVs were hard coded.

The changes of this PR provide a flexible schema-based system for defining the calibration state.  The new system continues to support legacy states without modification -- producing the same hash values and derived PV-values.  However, in addition the new system now allows the specification of the set of PVs contributing to a given state; any rounding of the PVs that should be applied during the hash computation; and a generalized representation of any derived PV-values required by the state.

This PR includes the following changes:

  * `StateId` has now been subsumed into `DetectorState`: this single class now retains detector-state related PVs and the associated hash-computation machinery.  Legacy `DetectorState` JSON format can also be successfully read by the new system.  (`StateId` itself had never been serialized so it has no legacy support.)

  * `InstrumentConfig` now retains the actual schema used to define the calibration state.  This schema is fully generalized in JSON-schema format, and allows any of the known PV-log types to be included in a state definition.  Rounding details affecting the hash computation, and derived-PV details are also included in this schema.  Legacy `InstrumentConfig` JSON format can be read by the new system.  In this legacy case, the `stateIdSchema` attribute defaults to `DetectorState.LEGACY_SCHEMA`.

Other related defects fixed:

  * The computation of compatible states, which are states sharing the same detector positions, previously used the unrounded PV values.  This has been corrected so that the rounded values, i.e. the same values used in the hash computation, are used to determine compatibility.

Known issues:

  * At present, the state-ID SHA is an attribute of `DetectorState`.  However, if strictly enforced, this causes problems with the current use of `StateId`-based magic numbers to store cached data.  For this reason, the existing `generateStateId` still exists, and returns `<detector state>.SHA, <detector state> | None`.  Of related interest, note that the previous method `stateIdFromDetectorState` has been removed.